requirements.txt
attrs==23.1.0
bidict==0.22.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
distlib==0.3.7
filelock==3.13.1
Flask==3.0.0
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
Pillow==10.1.0
platformdirs==3.11.0
python-engineio==4.8.0
python-socketio==5.10.0
simple-websocket==1.0.0
SQLAlchemy==2.0.23
timeago==1.0.16
typing_extensions==4.8.0
virtualenv==20.24.6
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.2
